Issue 65: LzoJsonRecordReader and JsonLoader may fail with NullPointerException
This patch prevents LzoJsonRecordReader and JsonLoader from failing with
a NPE when JSONParser#parse() returns a null value.  In this case a warning
is logged and the values 'false' (LzoJsonRecordReader) and 'null'
(JsonLoader) are returned, respectively.

JsonLoader#parseStringToTuple() will also increment the counter
JsonLoaderCounters.LinesParseError if this happens.

More information is available at:
https://github.com/kevinweil/elephant-bird/issues/65