Fix compatibility issue with pig9, upgrade to pig9
Pig9 changed the construction of bags, which was creating an issue in
the case of lists and sets with structs as elements. I added a test for
this case, and made it such that the change is consistent with both pig8
and pig9.

Note: the test will fail on pig8.