Add Ivy support to the build process. This was prompted by needing to generate a POM file so Elephant-Bird can be published to a Maven repository.
Remove most checked-in jars and download them with Ivy instead. Some jars are still checked in as they are not available in Ivy or I had issues getting the correct version for some reason.

Remove checked-in codegen, which is not generated under ./examples/build/gen-java.

Update serde per interface change in 0.8.0.