Issue 146: TestInvoker#testSpeed() may fail on slow machines
This patch removes the testSpeed() method altogether from the code.

Background:
The "problem" with testSpeed() is that its success seems to depend on
the build environment. Even though a relative ratio is used, this alone
apparently does not guarantee that the test completes successfully on
any box that the build is being run on (e.g. it may pass if the build
box is otherwise idle but it may fail if there are other tasks running
in the background). Also, this unit test is not a functional test but a
performance test.

See https://github.com/kevinweil/elephant-bird/issues/146 for more
information.