Update LuceneHdfsDirectory.java
When using inside a webapp, there were a lot of open files in (CLOSED_WAIT state) for the index files on HDFS for every request. A lot of open files eventually chokes the host. Also, the FileSystem should not be closed as it is passed from outside and it could be used by others.